redact-pdf/redact-pdf/requirements.txt
fastapi
uvicorn
pymupdf
pytesseract
pdf2image
Pillow
python-multipart
boto3
pydantic